refactor Bruin language server by removing BruinCompletionsWithCommands, consolidating completion logic, and enhancing asset data handling
